[DX] Extract ItemStep
