+++ spec/Pim/Bundle/BaseConnectorBundle/Archiver/FileReaderArchiverSpec.php
@@ -7,7 +7,7 @@ use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
 use Akeneo\Component\Batch\Item\ItemReaderInterface;
 use Akeneo\Bundle\BatchBundle\Job\Job;
 use Akeneo\Component\Batch\Step\AbstractStep;
-use Akeneo\Bundle\BatchBundle\Step\ItemStep;
+use Akeneo\Component\Batch\Step\ItemStep;
 use League\Flysystem\Filesystem;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\BaseConnectorBundle\Reader\File\CsvReader;
